{feature} introduce own harness interfaces
These interfaces should be compatible with Angular's, more or less, but
without pulling in any angular packages.

+++ scripts/ng-add/run.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-set -e
+set -ex
 
 DIR="$(mktemp -d)"
 cd "$DIR"
@@ -18,8 +18,19 @@ fi
 # add our package
 yarn ng add --registry "$(npm config get registry)" --skip-confirmation --defaults @ngx-playwright/test "$@"
 
+if [ -f playwright/tsconfig.json ]; then
+	yarn tsc --noEmit -p playwright/tsconfig.json
+fi
+
 # run the tests
 yarn ng e2e
+
+# Playwright has the annoying tendency to succeed if there are no tests, and
+# certain errors cause it to detect no tests without detecting the errors themselves
+if [ ! -f playwright/results/junit.xml ]; then
+	echo "Didn't find test result file, did the tests run?" >&2
+	exit 1
+fi
 
 # clean up disk space used by the test
 rm -rf "$DIR"
